Completed the Below Beta Feature:
Data entry screen changes:-
Make the following changes to the data entry screen:

1. After image selection, ask for "looking for"; user enters looking
for.
2. Ask for category; user selects category.
3. On the left and right corner of the screen, parallel to the "looking
for" line, display the occasion and find it at icons respectively. If
the user touches either of those, allow the user to enter the
information. Otherwise, do not prompt the user for this information.
4. After 2 is done, on the top action bar, provide 'cancel', 'next' and
'post' as 3 options (equal spacing between these options).
5. If the user chooses cancel, keep the current functionality on cancel
6. If the user chooses 'next', continue with prompting if they want to
add more images.
7. If the user chooses 'post', post the aisle without prompting the user
further.