Source selection changes :-
When the user chooses create aisle or add image, the source selection
screen appears. Make the following changes to it:

In the top menu, show -> camera, gallery, browser, more and 1 more app.
The one more app shown should be the first in this list that the user
has installed on their device:

1. Etsy
2. Fancy
3. Wanelo
4. Amazon
5. Pinterest
6. Zappos

If the user touches more, the rest of these apps should be shown in
order at the top (if installed on the device). At the bottom of the
list, there should be an option for "More" again, which will then list
all installed apps.

If none of these apps are on the user's phone, just display Etsy.
Touching it should give the option of going to the Play store, just as
it is today.